wip code reorganized
